trying to connect api with index
