+++ index.html
@@ -240,7 +240,8 @@
         </footer>
 
     </main>
-    <script src="/scrtips/index-page.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/axios@1.6.7/dist/axios.min.js"></script>
+    <script type="module" src="/scripts/index-page.js"></script>
 </body>
 
 </html>
